+++ blocks/header/header.js
@@ -144,27 +144,6 @@ function setActiveTab() {
   }
 }
 
-function buildCustomBreadcrumbs() {
-  const path = window.location.pathname.split('/').slice(1).map((word) => word.charAt(0).toUpperCase() + word.slice(1).replace(/-/g, ' '));
-
-  const breadcrumbContainer = document.createElement('div');
-  breadcrumbContainer.className = 'breadcrumb-container';
-
-  path.forEach((text, index) => {
-    const element = document.createElement(index === 0 ? 'a' : 'span');
-    if (index === 0) {
-      element.href = `/${text.toLowerCase().replace(/ /g, '-')}`;
-    }
-    element.textContent = text;
-    breadcrumbContainer.appendChild(element);
-    if (index < path.length - 1) {
-      breadcrumbContainer.appendChild(document.createTextNode(' • '));
-    }
-  });
-
-  document.querySelector('main').prepend(breadcrumbContainer);
-}
-
 /**
  * decorates the header, mainly the nav
  * @param {Element} block The header block element
@@ -314,9 +293,4 @@ export default async function decorate(block) {
 
   addAnimation();
   setActiveTab();
-
-  if (getMetadata('breadcrumbs').toLowerCase() === 'true') {
-    // navWrapper.append(await buildBreadcrumbs());
-    buildCustomBreadcrumbs();
-  }
 }
